+++ src/main/java/ru/ivmiit/service/BooksService.java
@@ -3,13 +3,13 @@ package ru.ivmiit.service;
 import ru.ivmiit.dto.BookDto;
 import ru.ivmiit.forms.BookForm;
 import ru.ivmiit.forms.EditBookStatusForm;
-import ru.ivmiit.model.Author;
 import ru.ivmiit.model.Book;
 
 import java.util.List;
 
 public interface BooksService {
     List<BookDto> getAllBooks();
+
 
     void addBook(BookForm form);
 
@@ -19,5 +19,5 @@ public interface BooksService {
 
     List<Book> getBooksByGenre(String genre);
 
-    Book getBookByAuthor(Integer id);
+    Book getBookByAuthor(Long id);
 }
